task for deploying
Initial task accepts path of deployment script runs it
