--- myjsonfile.json
@@ -1,1 +1,0 @@
-"{\"address\":\"d41b69c0b370fb6a17cdd3089aa4c42785dc38b6\",\"id\":\"3a6554d0-5c6a-4508-8598-fcd0d0ba8b6d\",\"version\":3,\"Crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"db32cfff06fd10cc42805a6ca8de2b8d\"},\"ciphertext\":\"fd4e136c9f44366d815d7f3d3b941c3b5d4acadd9d128dbf5deb68def2b7bb56\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"45744763a6e162b1a2c356bbf1027a4ffda00ed01a35733dc3a8d01760112c4b\",\"n\":131072,\"dklen\":32,\"p\":1,\"r\":8},\"mac\":\"dc8932c7092ded060433c52cc8a4e1ff8cd0839c8ab7448384d22caccfadb974\"},\"x-ethers\":{\"client\":\"ethers/6.7.0\",\"gethFilename\":\"UTC--2023-08-09T09-46-33.0Z--d41b69c0b370fb6a17cdd3089aa4c42785dc38b6\",\"path\":\"m/44'/60'/0'/0/0\",\"locale\":\"en\",\"mnemonicCounter\":\"f3c8846fd61532ba2687e82ce8528c2b\",\"mnemonicCiphertext\":\"549e1fa69c183e0f3ff8befc694cb34e\",\"version\":\"0.1\"}}"
