+++ scripts/deploy.js
@@ -7,6 +7,7 @@
 const hre = require("hardhat");
 
 async function main() {
+  console.log("Hello");
   const currentTimestampInSeconds = Math.round(Date.now() / 1000);
   const unlockTime = currentTimestampInSeconds + 60;
 
@@ -19,7 +20,7 @@ async function main() {
   await lock.waitForDeployment();
 
   console.log(
-    `Lock with ${ethers.formatEther(
+    `Lock with ${hre.ethers.formatEther(
       lockedAmount
     )}ETH and unlock timestamp ${unlockTime} deployed to ${lock.target}`
   );
@@ -28,6 +29,7 @@ async function main() {
 // We recommend this pattern to be able to use async/await everywhere
 // and properly handle errors.
 main().catch((error) => {
+  console.log("Errror");
   console.error(error);
   process.exitCode = 1;
 });
