--- superdeployer.json
@@ -1,1 +1,0 @@
-{"address":"ca7b22a32233c04b77f683c9117d2d0cc083ee1f","id":"1b6608bc-88f3-4b05-9320-dc58e964af0c","version":3,"Crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"e92ad8249be41548b13f14fd83c0e4cd"},"ciphertext":"4a7613ed4708229c5175e1fd312ce138a4fde66859e39f2f27a4bc46ef88bee4","kdf":"scrypt","kdfparams":{"salt":"e0914a1fe2f7160ea71d3115701cabcb3c9cee544eae58c1e60733405cc5ff93","n":131072,"dklen":32,"p":1,"r":8},"mac":"7827eed6bf169e8e7372f835c67bf2ecfa979dfafae7d11537061328c07b9095"},"x-ethers":{"client":"ethers/6.7.0","gethFilename":"UTC--2023-08-09T09-48-30.0Z--ca7b22a32233c04b77f683c9117d2d0cc083ee1f","path":"m/44'/60'/0'/0/0","locale":"en","mnemonicCounter":"db67fc01a2ae38f703f9d185395bc9b6","mnemonicCiphertext":"42219802546fd35ce48d87f70f52e660","version":"0.1"}}
